+++ .eslintrc.js
@@ -16,7 +16,7 @@ module.exports = {
     'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     "@typescript-eslint/ban-ts-comment": "off",
     "@typescript-eslint/no-inferrable-types": "off",
-    "@typescript-eslint/no-explicit-any": ["warn"],
+    "@typescript-eslint/no-explicit-any": ["off"],
   },
   overrides: [
     {