+++ packages/FormAuto/FormAuto.vue
@@ -18,11 +18,11 @@
 					v-for="(item, name) in fields"
 					:is="inline ? 'span' : 'el-col'"
 					:span="item.col || 24"
+					:class="{'el-form-item_hidden':item.type=='hidden'}"
 					:key="`col_${name}`"
 				>
 					<el-form-item
 						v-if="!item.bindShow || item.bindShow(model)"
-						:class="{'el-form_hidden':item.type=='hidden'}"
 						:prop="name"
 						:label-width="item.labelWidth"
 						:key="`formItem_${name}`"
@@ -43,13 +43,7 @@
 							<div v-bind="item.props">{{ model[name] }}</div>
 						</template>
 						<template v-else-if="/text|password|textarea/g.test(item.type)">
-							<el-input
-								v-model="model[name]"
-								:type="item.type"
-								clearable
-								v-bind="item.props"
-								v-on="item.on"
-							></el-input>
+							<el-input v-model="model[name]" :type="item.type" v-bind="item.props" v-on="item.on"></el-input>
 						</template>
 						<template v-else-if="item.type=='hidden'">
 							<el-input v-model="model[name]" type="hidden" v-bind="item.props"></el-input>
@@ -144,7 +138,7 @@
 							</el-checkbox-group>
 						</template>
 						<template v-else-if="item.type == 'select'">
-							<el-select v-model="model[name]" clearable v-bind="item.props" v-on="item.on">
+							<el-select v-model="model[name]" v-bind="item.props" v-on="item.on">
 								<el-option
 									v-for="(option, key) in item.options"
 									:key="`${name}_${key}`"
@@ -160,7 +154,6 @@
 							<el-cascader
 								v-model="model[name]"
 								:options="item.options"
-								clearable
 								v-bind="item.props"
 								v-on="item.on"
 							></el-cascader>
@@ -200,7 +193,7 @@ import {
 import { Form } from "element-ui";
 import { debounce, cloneDeep, forEach, omit } from "lodash";
 import { ElFormAutoField } from "../../types/form-auto";
-import { ElAutoOption } from "../../types/saas-extend"
+import { ElAutoMixinOptions, ElAutoOption } from "../../types/saas-extend"
 // @ts-ignore
 import DynamicSlot from "../components/DynamicSlot.js"
 import { transformOptions } from "../util"
@@ -262,7 +255,6 @@ export default class ElFormAuto extends Vue {
 
 	@Watch("value", { immediate: true, deep: true })
 	private onValueChange(value: Record<string, any>) {
-		console.log(value)
 		value && this.setModel(value);
 	}
 
@@ -329,7 +321,6 @@ export default class ElFormAuto extends Vue {
 	 * @param {object} model 表单项对应值数据 例如：{key:value,...}
 	 */
 	public setModel(model: Record<string, any>): void {
-		console.log(this.value)
 		if (!this.isRender) return
 		let _model = Object.assign({}, model);
 		for (let name in _model) {
@@ -463,9 +454,14 @@ export default class ElFormAuto extends Vue {
 				}
 			}
 
+			if (/text|password|textarea|select|cascader/.test(item.type)) {
+				item.props.clearable = item.props.clearable == false ? false : true
+			}
+
 			if (/select|radio|check/.test(item.type)) {
 				this.asyncOptions.push(item)
 			}
+
 			if (this.isRender == false) {
 				this.$set(this.model, name, this.value[name] || item.value);
 				if (item.type == "check" && item.checkAll !== false) {
@@ -522,7 +518,7 @@ export default class ElFormAuto extends Vue {
 						item.props.remoteMethod = debounce((query?: string) => {
 							item.options = []
 							item.props && (item.props.loading = true);
-							remoteMethod(query).then((options: any) => {
+							remoteMethod(query).then((options: ElAutoMixinOptions) => {
 								return transformOptions(options)
 							}).then((options: any) => {
 								item.options = options;