+++ types/form-auto.d.ts
@@ -97,7 +97,7 @@ export declare interface ElFormAutoField {
    * 选项数组，支持Promise
    * @default false
    */
-  options?: ElAutoMixinOptions | ((query?: string) => any);
+  options?: ElAutoMixinOptions | ((query?: string) => Promise<ElAutoMixinOptions>);
   /**
    * type为check时
    * 去除 全选复选框